+++ src/test/kotlin/db/tables/BlockedUsersTest.kt
@@ -13,6 +13,8 @@ import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.extension.ExtendWith
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFalse
+import kotlin.test.assertTrue
 
 @ExtendWith(DbExtension::class)
 class BlockedUsersTest {
@@ -90,7 +92,7 @@ class BlockedUsersTest {
         fun `The user must be unblocked`() {
             val (blockerId, blockedId) = createVerifiedUsers(2).map { it.info.id }
             BlockedUsers.create(blockerId, blockedId)
-            BlockedUsers.delete(blockerId, blockedId)
+            assertTrue(BlockedUsers.delete(blockerId, blockedId))
             assertEquals(0, BlockedUsers.count())
         }
 
@@ -101,7 +103,7 @@ class BlockedUsersTest {
             awaitBrokering()
             val (blockerSubscriber, blockedSubscriber) =
                 listOf(blockerId, blockedId).map { accountsNotifier.subscribe(it).subscribeWith(TestSubscriber()) }
-            BlockedUsers.delete(blockerId, blockedId)
+            assertTrue(BlockedUsers.delete(blockerId, blockedId))
             awaitBrokering()
             blockerSubscriber.assertValue(UnblockedAccount(blockedId))
             blockedSubscriber.assertNoValues()
@@ -111,7 +113,7 @@ class BlockedUsersTest {
         fun `The user mustn't be notified when unblocking a user who wasn't blocked`(): Unit = runBlocking {
             val (blockerId, blockedId) = createVerifiedUsers(2).map { it.info.id }
             val subscriber = accountsNotifier.subscribe(blockerId).subscribeWith(TestSubscriber())
-            BlockedUsers.delete(blockerId, blockedId)
+            assertFalse(BlockedUsers.delete(blockerId, blockedId))
             awaitBrokering()
             subscriber.assertNoValues()
         }